\name{Comment}
\alias{Comment}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Add a comment to a package
}
\description{
Make a comment on a package using less than 1000 characters.
}
\usage{
Comment(package, comment, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{package}{
The name of the package to comment on
}
  \item{comment}{
Comments
}
  \item{\dots}{
Argument can be passed to the function find.package.
}
}
