\name{dislike}
\alias{dislike}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Vote to dislike a package
}
\description{
Vote to dislike a package
}
\usage{
dislike(package, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{package}{
The name of the package
}
  \item{\dots}{
Argument can be passed to the function find.package.
}
}
