\name{like}
\alias{like}
\title{
Vote to like a package
}
\description{
Vote to like a package
}
\usage{
like(package, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{package}{
The name of the package
}
  \item{\dots}{
Argument can be passed to the function find.package.
}
}