% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_percent.r
\name{parenthesize}
\alias{parenthesize}
\title{Convert a 3-vector to range with parenthesis, i.e., 1(2-3) '}
\usage{
parenthesize(r)
}
\arguments{
\item{r}{input vector of length 3}
}
\description{
This is a basic function
}
