\name{rate}
\alias{rate}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Rate an package using scores 1-5
}
\description{
Rate an package using scores 1-5
}
\usage{
rate(package, rating, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{package}{
The name of the package
}
  \item{rating}{
Rating of 1, 2, 3, 4, 5 from Bad, OK, Good, Very Good, Excellent
}
  \item{\dots}{
Argument can be passed to the function find.package.
}
}
