% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_percent.r
\name{to_percent}
\alias{to_percent}
\title{Convert float to percentage (text)'}
\usage{
to_percent(x, digits = 4)
}
\arguments{
\item{x}{input float (e.g., .78)}

\item{digits}{number of digits (4 is usual}
}
\description{
This is a basic function
}
