\name{download}
\alias{download}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Tell that you download the package
}
\description{
Tell that you download the package
}
\usage{
download(package, ...)
}
\arguments{
  \item{package}{
The name of the package
}
  \item{\dots}{
Argument can be passed to the function find.package.
}
}