\name{view}
\alias{view}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
View the usage information about an package
}
\description{
View the usage information about an package
}
\usage{
view(package, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{package}{
The name of the package
}
  \item{\dots}{
Argument can be passed to the function find.package.
}
}
\details{
The usage information includes number of downloads, number of likes, number of dislikes, and the average rating.
}
